+++ src/bda/plone/orders/interfaces.py
@@ -136,6 +136,18 @@ class IBookingCancelledEvent(Interface):
     booking_attrs = Attribute(u"Dict of attributes of the cancelled booking.")
 
 
+class IBookingReservedToOrderedEvent(Interface):
+    """Checkout related event.
+    """
+    context = Attribute(u"Context in which this event was triggered.")
+
+    request = Attribute(u"Current request.")
+
+    order_uid = Attribute(u"UUID of Order")
+
+    booking_attrs = Attribute(u"Dict of attributes of the cancelled booking.")
+
+
 class IStockThresholdReached(Interface):
     """Checkout related event.
     """